+++ src/contexts/remoteConfigs.tsx
@@ -14,12 +14,15 @@ export interface Configs {
   premium_max_participants: string;
   default_max_groups: string;
   premium_max_groups: string;
+  default_max_message_length: string;
+  premium_max_message_length: string;
 }
 
 interface UserConfigs {
   fileUpload: number;
   amountGroups: number;
   amountParticipants: number;
+  messageLength: number;
 }
 
 const RemoteConfigsContext = createContext({} as RemoteConfigContextProps)
@@ -56,14 +59,16 @@ const RemoteConfigsProvider: React.FC = ({ children }) => {
       return setUserConfigs({
         fileUpload: parseFloat(allConfigs.premium_file_upload),
         amountGroups: parseInt(allConfigs.premium_max_groups),
-        amountParticipants: parseInt(allConfigs.premium_max_participants)
+        amountParticipants: parseInt(allConfigs.premium_max_participants),
+        messageLength: parseInt(allConfigs.premium_max_message_length)
       })
     }
 
     return setUserConfigs({
       fileUpload: parseFloat(allConfigs.default_file_upload),
       amountGroups: parseInt(allConfigs.default_max_groups),
-      amountParticipants: parseInt(allConfigs.default_max_participants)
+      amountParticipants: parseInt(allConfigs.default_max_participants),
+      messageLength: parseInt(allConfigs.default_max_message_length)
     })
   }